+++ CHANGELOG.md
@@ -13,3 +13,11 @@
 ## 0.0.7+1
 
 * Change encode image.
+
+## 0.0.7+2
+
+* Add file or Uint8List to convert type.
+
+## 0.0.8
+
+* Add video/image preview & add some tools.